connections and actionProcessor moved to classes
