+++ bin/templates/server.js
@@ -30,7 +30,9 @@ module.exports = class MyServer extends ActionHero.Server {
   }
 
   start () {
-
+    // this.buildConnection (api, data)
+    // this.processAction (api, connection)
+    // this.processFile (connection)
   }
 
   stop () {
