+++ index.js
@@ -1,8 +1,15 @@
-const path = require('path')
+const path = require('path');
 
-exports.Process = require(path.join(__dirname, 'classes', 'process.js'))
-exports.Action = require(path.join(__dirname, 'classes', 'action.js'))
-exports.Task = require(path.join(__dirname, 'classes', 'task.js'))
-// exports.Initializer = require(path.join(__dirname, 'classes', 'initializer.js'))
-exports.Server = require(path.join(__dirname, 'classes', 'server.js'))
-exports.CLI = require(path.join(__dirname, 'classes', 'cli.js'))
+[
+  'Process',
+  'Action',
+  'Task',
+  'Initializer',
+  'Server',
+  'CLI',
+  'ActionProcessor',
+  'Connection'
+].forEach((component) => {
+  let fileName = component.charAt(0).toLowerCase() + component.substring(1)
+  exports[component] = require(path.join(__dirname, 'classes', `${fileName}.js`))
+})
